refactor: embedded 'affelnet_infos_offre' form in publish modal
and extract publish modal in dedicated component
